+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # unreleased
 
+- Adds Minitest::Suite.order= for specifying initial suite order
 - Adds MINITEST_SUITE_ONLY and MINITEST_SUITE_EXCEPT options
 
 # 0.0.1